reorgnizing the project into resourse & main code
